pyproject.toml
[project]
name = "ai-closer"
version = "0.1.0"
description = "AI Sales Closer for Telegram Stores — multi-tenant SaaS platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "telethon>=1.36.0",
    "openai>=1.30.0",
    "celery[redis]>=5.4.0",
    "python-multipart>=0.0.9",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
